Extracted isHTMLValid function to TestHelpers directory
