+++ test/SimpleTests.test.js
@@ -1,33 +1,13 @@
 const renderer = require("../index")
 const exampleCVJSON = require("@jsonresume/schema/sample.resume.json")
 const fs = require('node:fs');
-const HtmlValidate = require("html-validate");
+const {isHTMLValid} = require("./TestHelpers/HTMLValidate");
 
 
 const testOutPutPath = `${__dirname}/TestOutput`;
 
-const isHTMLValid = async (html, logResults = true) => {
-  const htmlvalidate = new HtmlValidate.HtmlValidate({
-    extends: ["html-validate:recommended"],
-    rules: {
-      "doctype-style": "off",
-      "attr-quotes": "off",
-      "no-trailing-whitespace": "off",
-      "void-style": "warn",
-    },
-  });
 
-  // See https://html-validate.org/guide/api/getting-started.html
-  const validationReport = await htmlvalidate.validateString(html);
-
-  if (logResults) {
-    validationReport.results.forEach(console.log);
-  }
-
-  return validationReport.valid;
-}
-
-describe("SimpleTests",  () => {
+describe("SimpleTests", () => {
 
   var result;
